A lot of very small changes. Most important is prparing of a more sofisticated method to determine the bandwidth and the MP3 sampling frequency.

+++ frontend/get_audio.h
@@ -61,7 +61,7 @@ int lame_decoder(lame_global_flags *gfp,FILE *outf,int skip, char *inPath, char 
 enum byte_order { order_unknown, order_bigEndian, order_littleEndian };
 extern enum byte_order NativeByteOrder;
 enum byte_order DetermineByteOrder(void);
-void SwapBytesInWords( short *loc, int words );
+void SwapBytesInWords( short *loc, size_t words );
